requirements.txt
Django==4.0.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.12.0
